+++ src/blocks/homepage-articles/edit.js
@@ -434,6 +434,16 @@ class Edit extends Component {
 							/>
 						)
 					) }
+					<ToggleControl
+						label={ __( 'Use deduplication logic', 'newspack-blocks' ) }
+						help={ __(
+							'If unchecked, this block will be excluded from the deduplication logic and may show duplicate posts.',
+							'newspack-blocks'
+						) }
+						checked={ attributes.deduplicate }
+						onChange={ () => setAttributes( { deduplicate: ! attributes.deduplicate } ) }
+						className="newspack-blocks-deduplication-toggle"
+					/>
 				</PanelBody>
 				<PanelBody title={ __( 'Featured Image Settings', 'newspack-blocks' ) }>
 					<PanelRow>